packages/cli: moved build cache implementation to lib
